% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{get_ci}
\alias{get_ci}
\title{Get the CI for a Multinomial distribution}
\usage{
get_ci(x, method = "sisonglaz", conf.level = 0.95)
}
\arguments{
\item{x}{
A vector of positive integers representing the number of occurrences of each class. The total number of samples equals the sum of such elements.
}

\item{method}{ character string specifing which method to use; can be one out of
  \code{"sisonglaz"}, \code{"cplus1"}, \code{"goodman"},  \code{"wald"},  \code{"waldcc"}, \code{"wilson"}.
  Method can be abbreviated. See details. Defaults to \code{"sisonglaz"}. }

\item{conf.level}{ confidence level, defaults to 0.95. }
}
\value{
A data frame with confidence intervals calculated.
}
\description{
A wrapper for \link[DescTools]{MultinomCI} for two-sided confidence intervals of the Multinomial distribution.
}
