% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{is.rpsi}
\alias{is.rpsi}
\title{rpsi object}
\usage{
is.rpsi(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
Boolean. \code{TRUE} when \code{x} is of class \code{rpsi}.
}
\description{
An internal generic function. Methods for \code{rpsi} should only return \code{TRUE} if the class is \code{rpsi}.
}
