% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{method_names}
\alias{method_names}
\title{Get formatted method names}
\usage{
method_names(x)
}
\arguments{
\item{x}{Method name from \code{method} argument of \link[DescTools]{MultinomCI}.}
}
\value{
A formatted string.
}
\description{
Get formatted method names
}
