% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.rpsi}
\alias{plot.rpsi}
\title{Plot rpsi object}
\usage{
\method{plot}{rpsi}(x, crit_vals = c(0.95, 0.99))
}
\arguments{
\item{x}{An object of class \code{rpsi}.}

\item{...}{Redundant argument for consistency with method.}
}
\value{
No return value.
}
\description{
Plot rpsi object
}
