% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{PSI-package}
\alias{PSI}
\alias{PSI-package}
\title{Calculating the Population Stability Index (PSI)}
\description{
\pkg{PSI} allows you to to calculate the population stability index in an
easy manner.

This package only contains one function \code{\link{psi}}, which you could
learn how to use it from the document.
}
\author{
\strong{Maintainer}: Siyuan Yao \email{siyuan.yao@gmail.com}

}
