% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{psi}
\alias{psi}
\title{Calculate PSI metric}
\usage{
psi(x, y, var, count, random_base = TRUE)
}
\arguments{
\item{x}{A data frame containing the distinct groups and counts of the base distribution. Must contain columns \code{var} and \code{count}.}

\item{y}{The comparison distribution in the same format as \code{x}.}

\item{var}{The name of the column in \code{x} and \code{y} representing the discrete groups of the distributions.}

\item{count}{The name of the column in \code{x} and \code{y} representing the count of observations in each group.}

\item{random_base}{A logical value indicating whether the base distribution uesd for comparison is considered random or if it is considered a population.}
}
\value{
An object of class \code{rpsi}. See \link[rpsi]{rpsi} for details.
}
\description{
Calculates the PSI metric with thresholds calculated from Yurdakul, Bilal (2018). \loadmathjax
}
\details{
See \href{https://scholarworks.wmich.edu/cgi/viewcontent.cgi?article=4249&context=dissertations}{Yurdakul, Bilal (2018)} for details.

The PSI is shown to be approximately distributed as
\deqn{PSI\sim\chi^{2}}
}
\examples{
library(rpsi)
library(dplyr)
x = data.frame(x = rnorm(500, 0.1)) \%>\% mutate(VAL  = factor(ifelse(x< -2,"A", ifelse(x<0, "B", ifelse(x>2, "D", "C"))))) \%>\% group_by(VAL) \%>\% summarise(N = n())
y = data.frame(x = rnorm(500, 0.1)) \%>\% mutate(VAL  = factor(ifelse(x< -2,"A", ifelse(x<0, "B", ifelse(x>2, "D", "C"))))) \%>\% group_by(VAL) \%>\% summarise(N = n())
psi(x, y, var = "VAL", count = "N")
}
