% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rpsi_pal}
\alias{rpsi_pal}
\title{Define an `rpsi` colour palette}
\usage{
rpsi_pal(palette = "blues", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Must be a color from \link[rpsi]{rpsi_cols}.}

\item{reverse}{Whether to reverse the palette direction.}

\item{...}{Other arguments passed onto \link[grDevices]{colorRampPalette}.}
}
\description{
Define an `rpsi` colour palette
}
