% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.rpsi}
\alias{summary.rpsi}
\title{Summarise rpsi object}
\usage{
\method{summary}{rpsi}(object, crit_val = 0.99, ...)
}
\arguments{
\item{object}{An object of class \code{rpsi}.}

\item{crit_val}{The significance level to use.}

\item{...}{Redundant argument for consistency with method.}
}
\value{
No return value.
}
\description{
Summarise rpsi object
}
