% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.rpsi}
\alias{plot.rpsi}
\title{Plot rpsi object}
\usage{
\method{plot}{rpsi}(x, crit_val = 0.99, fill.col = "blues", ...)
}
\arguments{
\item{x}{An object of class \code{pkgdepR}.}

\item{crit_val}{The significance level to use.}

\item{fill.col}{The colour palette to use. Must be a palette from \link[rpsi]{rpsi_cols}().}

\item{...}{Redundant argument for consistency with method.}
}
\value{
A list of object of classes \code{gg} and \code{ggplot}.
}
\description{
This function
}
