% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.rpsi}
\alias{print.rpsi}
\title{Print rpsi object}
\usage{
\method{print}{rpsi}(x, crit_val = 0.99, ...)
}
\arguments{
\item{x}{An object of class \code{rpsi}.}

\item{crit_val}{The significance level to use.}

\item{...}{Redundant argument for consistency with method.}
}
\value{
No return value.
}
\description{
Print rpsi object
}
