% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Population Stability Index}
\usage{
psi(original, current, cut.points = NULL)
}
\arguments{
\item{original}{The original set of a measurement, should be a factor or numeric}

\item{current}{The current set of a measurement, should be a factor or numeric}

\item{cut.points}{It won't work if original and current are factors, and it cannot be NULL if original and current are numerical. This functoin use this argument to binning original and current with left-closed right-open interval.}
}
\description{
Population Stability Index
}
